Moved valid check into error type computation
